✨ migrate chain store to pinia
