+++ components/accounts/utils.ts
@@ -1,7 +1,7 @@
-import * as store from '~/store'
 import correctFormat from '~/utils/ss58Format'
 import { encodeAddress, isAddress } from '@polkadot/util-crypto'
-import { createInteraction, Interaction } from '@kodadot1/minimark'
+import { Interaction, createInteraction } from '@kodadot1/minimark'
+import { useChainStore } from '@/stores/chain'
 
 export type ShuffleFunction = (arr: string[]) => string[]
 export type ProcessFunction = (
@@ -21,7 +21,8 @@ export const toDistribute = (length: number, distribution: number): number => {
 }
 
 export const parseBatchAddresses = (batchAddresses: string): string[] => {
-  const ss58Format: any = store.getters['chain/getChainProperties58Format']
+  const chainStore = useChainStore()
+  const ss58Format = chainStore.getChainProperties58Format
   const addresses = batchAddresses
     .split('\n')
     .map((x) => x.split('-'))
