+++ plugins/polkadot.ts
@@ -1,9 +1,11 @@
 import keyring from '@polkadot/ui-keyring'
 import { cryptoWaitReady } from '@polkadot/util-crypto'
 import correctFormat from '~/utils/ss58Format'
+import { useChainStore } from '@/stores/chain'
 
-export default ({ store, isDev }): void => {
-  const ss58Format = store.getters['chain/getChainProperties58Format']
+export default ({ isDev }): void => {
+  const chainStore = useChainStore()
+  const ss58Format = chainStore.getChainProperties58Format
   cryptoWaitReady().then(() => {
     keyring.loadAll({
       ss58Format: correctFormat(ss58Format),
